Consolidate registration into AuthController and add validation
